+++ CHANGELOG.md
@@ -20,8 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * `jj git push` will search `@-` for branches to push if `@` has none.
 
-* The new revset function `file(pattern)` finds commits modifying the
-  paths specified by the `pattern`.
+* The new revset function `file(pattern..)` finds commits modifying the
+  paths specified by the `pattern..`.
 
 ### Fixed bugs